requirements.txt
fastapi
google-cloud-bigquery
google-cloud-pubsub
uvicorn
scikit-learn
gunicorn
redis
PyYAML==6.0.1
